refactor: separate server from client interface
